+++ frontend/src/pages/ProjectDetail/ProjectDetailPage.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect } from 'react';
+import React, { useEffect, useState } from 'react';
 import { useParams, useHistory } from 'react-router-dom';
 import {
   IonPage,
@@ -31,20 +31,53 @@ import {
   peopleCircleOutline,
 } from 'ionicons/icons';
 import JoinProjectButton from '../../components/JoinProjectButton/JoinProjectButton';
+import { getExistingToken } from '../../utils/authUtils';
+import { deleteParticipant, postParticipant, ResultType } from '../../utils/globalDataUtils';
+import { useToast } from '../../components/ToastProvider';
 
 interface ProjectDetailPageProps {
   profile: Profile | null;
   event: Event | null;
   projects: Project[];
+  onParticipantChange: () => void;
 }
 
-const ProjectDetailPage: React.FC<ProjectDetailPageProps> = ({ profile, event, projects }) => {
+const ProjectDetailPage: React.FC<ProjectDetailPageProps> = ({ profile, event, projects, onParticipantChange }) => {
   const { id } = useParams<{ id: string }>();
   const history = useHistory();
+  const [loading, setLoading] = useState<boolean>(false);
+  const { showToastMessage, showToastError } = useToast();
+
   const project = projects.find((p) => p.id === parseInt(id));
 
-  const handleJoinProject = (id: number) => {
-    console.log(`Projekt mit ID ${id} beitreten`);
+  const handleJoinProject = async (project_id: number, user_id: number) => {
+    console.log(`Projekt mit ID ${project_id} beitreten`);
+    setLoading(true);
+    const token = getExistingToken();
+    const result = await postParticipant(project_id, user_id, token);
+
+    if (result.resultType !== ResultType.SUCCESS || result.data === null) {
+      showToastError(result.resultMsg ?? 'Fehler beim Speichern des Projekts');
+    } else {
+      showToastMessage('Projekt erfolgreich beigetreten!');
+    }
+    onParticipantChange();
+    setLoading(false);
+  };
+
+  const handleRejectProject = async (project_id: number, user_id: number) => {
+    console.log(`Projekt mit ID ${project_id} ablehnen`);
+    setLoading(true);
+    const token = getExistingToken();
+    const result = await deleteParticipant(project_id, user_id, token);
+
+    if (result.resultType !== ResultType.SUCCESS || result.data === null) {
+      showToastError(result.resultMsg ?? 'Fehler beim Löschen der Teilnahme');
+    } else {
+      showToastMessage('Teilnahme erfolgreich zurückgezogen!');
+    }
+    onParticipantChange();
+    setLoading(false);
   };
 
   const handleEditProject = () => {
@@ -68,23 +101,23 @@ const ProjectDetailPage: React.FC<ProjectDetailPageProps> = ({ profile, event, p
     );
   }
 
-  const isInitiator = project.initiators.some((initiator) => initiator.id === profile?.id);
+  const isInitiator   = project.initiators.some((initiator) => initiator.id === profile?.id);
+  const isParticipant = project.participants?.some((p) => p.id === profile?.id);
 
   return (
     <IonPage>
       <IonContent>
-        {isInitiator && (
-          <IonFab vertical="top" horizontal="end" slot="fixed">
-            <IonFabButton color="primary" onClick={handleEditProject}>
-              <IonIcon icon={pencilOutline} />
-            </IonFabButton>
-          </IonFab>
-        )}
         <IonGrid>
           <IonRow>
             {/* Projekt Details */}
             <IonCol size="12" sizeLg="8">
               <IonCard className="hackathon-card project-detail-card">
+                <div className="project-badge-container">
+                  {isInitiator && <div className="project-badge initiator-badge">Initiator</div>}
+                  {!isInitiator && isParticipant && (
+                    <div className="project-badge participant-badge">Teilnehmer</div>
+                  )}
+                </div>
                 <IonCardHeader>
                   <IonCardTitle style={{ fontSize: '1.5rem', fontWeight: 'bold' }}>
                     {project.idea}
@@ -134,6 +167,13 @@ const ProjectDetailPage: React.FC<ProjectDetailPageProps> = ({ profile, event, p
                       </IonLabel>
                     </IonItem>
                   </IonList>
+                  {isInitiator && (
+                    <IonFab vertical="bottom" horizontal="end" slot="fixed" className="fab-inside-card">
+                      <IonFabButton color="primary" onClick={handleEditProject}>
+                        <IonIcon icon={pencilOutline} />
+                      </IonFabButton>
+                    </IonFab>
+                  )}
                 </IonCardContent>
               </IonCard>
             </IonCol>
@@ -165,7 +205,9 @@ const ProjectDetailPage: React.FC<ProjectDetailPageProps> = ({ profile, event, p
                     <JoinProjectButton
                       project={project}
                       profile={profile!}
-                      onJoinProject={() => handleJoinProject(project.id)}
+                      onRejectProject={() => handleRejectProject(project.id, profile!.id)}
+                      onJoinProject={() => handleJoinProject(project.id, profile!.id)}
+                      disabled={loading}
                     />
                   </div>
                 </IonCardContent>